pgwire: move the parsing logic for client opts to a new file
Release note: None
